requirements.txt
Pillow==9.3.0
pygame==2.1.3.dev8
pytube==15.0.0
requests==2.28.1
regex==2022.10.31